Crate ctype

CType Pallet

A simple pallet which enables users to store their CType hash (blake2b as hex string) on chain and associate it with their account id.

Config

Call

Pallet

Terminology

CType:: CTypes are claim types. In everyday language, they are standardised structures for credentials. For example, a company may need a standard identification credential to identify workers that includes their full name, date of birth, access level and id number. Each of these are referred to as an attribute of a credential.

Assumptions

The CType hash was created using our KILT JS-SDK.

The underlying CType includes only the following required fields for the JSON-Schema we use in the SDK: Identifier, KILT specific JSON-Schema, Title and Properties.

Re-exports

pub use crate::default_weights::WeightInfo;

pub use crate::pallet::*;

Modules

benchmarking

ctype_entry

default_weights

Autogenerated weights for ctype

mock

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.